+++ HISTORY.rst
@@ -13,6 +13,7 @@ History
 - Take buttons in the file view out of the scroll area (nens/rana#3630)
 - Make creating avatars safe for users without given and/or family name
 - Download results folder "<simulation name> (<simulation id>)" (nens/rana#3678)
+- Add publications browser (nens/rana#302)
 
 
 1.2.6 (2026-02-19)